除最后的bonus全部已经完成

+++ src/NonUniformMesh.cpp
@@ -1,3 +1,4 @@
+// NonUniformMesh.cpp
 #include "NonUniformMesh.h"
 #include <cmath>
 
@@ -7,7 +8,7 @@ NonUniformMesh::NonUniformMesh(double x_min, double x_max, double dx)
     x_points.resize(n_points);
     for (int i = 0; i < n_points; ++i) {
         double ratio = static_cast<double>(i) / (n_points - 1);
-        x_points[i] = x_min + (x_max - x_min) * std::pow(ratio, 2); // 示例的非均匀分布
+        x_points[i] = x_min + (x_max - x_min) * std::pow(ratio, 2);
     }
 }
 
@@ -20,15 +21,15 @@ double NonUniformMesh::getFinalPosition() const {
 }
 
 int NonUniformMesh::getNumberOfPoints() const {
-    return x_points.size();
+    return (int)x_points.size();
 }
 
 double NonUniformMesh::getStepSize() const {
-    return dx; // 对于非均匀网格，这个值可能没有实际意义
+    return dx; 
 }
 
 int NonUniformMesh::x_size() const {
-    return x_points.size();
+    return (int)x_points.size();
 }
 
 double NonUniformMesh::x_size(int i) const {
